requirements.txt
streamlit
openai
pandas
python-dateutil
